+++ public/tagfiles.php
@@ -5,20 +5,38 @@ use diskover\Constants;
 error_reporting(E_ALL ^ E_NOTICE);
 require __DIR__ . "/../src/diskover/Diskover.php";
 
-// Update files if any were submitted
-if (count($_POST['ids']) > 0) {
-  // Connect to Elasticsearch
-  $client = connectES();
+// Connect to Elasticsearch
+$client = connectES();
 
-  // Update the files in Elasticsearch
-  foreach ($_POST['ids'] as $id => $value) {
+// Update files if any were submitted
+if (count($_POST['ids_tag']) > 0) {
+  
+  // update existing tag field with new value
+  foreach ($_POST['ids_tag'] as $id => $value) {
     $index = $_POST[$id];
     $params = array();
     $params['id'] = $id;
     $params['index'] = $index;
     $params['type'] = Constants::ES_TYPE;
     $result = $client->get($params);
-    $result['_source']['tag'] = $value; // update existing field with new value
+    $result['_source']['tag'] = $value;
+
+    $params['body']['doc'] = $result['_source'];
+    $result = $client->update($params);
+  }
+}
+
+if (count($_POST['ids_tag_custom']) > 0) {
+
+  // update existing tag_custom field with new value
+  foreach ($_POST['ids_tag_custom'] as $id => $value) {
+    $index = $_POST[$id];
+    $params = array();
+    $params['id'] = $id;
+    $params['index'] = $index;
+    $params['type'] = Constants::ES_TYPE;
+    $result = $client->get($params);
+    $result['_source']['tag_custom'] = $value;
 
     $params['body']['doc'] = $result['_source'];
     $result = $client->update($params);